+++ src/components/maps/FoundLocateUser.tsx
@@ -1,0 +1,61 @@
+import { useEffect, useState, useCallback } from "react";
+import { useMap } from "react-leaflet";
+
+interface LocateUserProps {
+  onLocationFound: (lat: number, lng: number) => void;
+  userCoords: { lat: number; lng: number } | null;
+}
+
+const LocateUser: React.FC<LocateUserProps> = ({ onLocationFound, userCoords }) => {
+  const map = useMap();
+  const [hasCentered, setHasCentered] = useState(false);
+  const [locationRequested, setLocationRequested] = useState(false);
+
+  const handleLocationFound = useCallback(
+    (e: L.LocationEvent) => {
+      const { lat, lng } = e.latlng;
+      console.log("Ubicación encontrada:", { lat, lng });
+      onLocationFound(lat, lng);
+    },
+    [onLocationFound]
+  );
+
+  const handleLocationError = useCallback((error: L.ErrorEvent) => {
+    console.error("Error al obtener la ubicación:", error.message);
+    alert("No se pudo obtener tu ubicación. Por favor, habilita la geolocalización.");
+  }, []);
+
+  // Centrar el mapa en la ubicación del usuario si está disponible
+  useEffect(() => {
+    if (userCoords && map && !hasCentered) {
+      console.log("Centrando el mapa en la ubicación del usuario:", userCoords);
+      map.setView([userCoords.lat, userCoords.lng], 15);
+      setHasCentered(true);
+    }
+  }, [userCoords, map, hasCentered]);
+
+  // Solicitar la ubicación solo una vez y manejar el watch
+  useEffect(() => {
+    if (!locationRequested && map) {
+      setLocationRequested(true);
+      map.locate({
+        setView: false,
+        maxZoom: 15,
+        watch: true,
+      });
+
+      map.on("locationfound", handleLocationFound);
+      map.on("locationerror", handleLocationError);
+
+      // Limpiar los event listeners cuando el componente se desmonte o la ubicación cambie
+      return () => {
+        map.off("locationfound", handleLocationFound);
+        map.off("locationerror", handleLocationError);
+      };
+    }
+  }, [map, locationRequested, handleLocationFound, handleLocationError]);
+
+  return null;
+};
+
+export default LocateUser;
